refactor(nfts): lazy load nft screen
