+++ packages/extension/src/ui/index.tsx
@@ -1,19 +1,8 @@
-import "@google/model-viewer/dist/model-viewer"
-
 import { StrictMode } from "react"
 import { createRoot } from "react-dom/client"
 import { BrowserRouter } from "react-router-dom"
 
 import { App } from "./App"
-
-/* eslint-disable @typescript-eslint/no-namespace */
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      "model-viewer": any
-    }
-  }
-}
 
 const container = document.getElementById("root")
